backend/requirements.txt
numpy
matplotlib
seaborn
scikit-learn
wordcloud
nltk
xgboost
streamlit
flask
flask-cors
plotly
pandas
openpyxl
xlrd
pymongo
python-dotenv
clerk-sdk-python
requests
PyJWT
cryptography
gunicorn
